refactor!: remove download_asset from Client
